+++ apps/web/config/routes.rb
@@ -1,4 +1,5 @@
-get "/", to: "dashboard#index",       as: :root
+get "/sync_processes/:id/stats", to: "sync_processes#stats", as: :stats
+get "/",                         to: "dashboard#index",      as: :root
 
 resources :errors,         only: [:index, :show], controller: "external_errors"
 resources :reservations,   only: [:index]