+++ apps/web/controllers/sync_processes/stats.rb
@@ -1,0 +1,14 @@
+module Web::Controllers::SyncProcesses
+  class Stats
+    include Web::Action
+    include Web::Controllers::InternalError
+
+    expose :sync_process
+
+    def call(params)
+      @sync_process = SyncProcessRepository.find(params[:id])
+
+      halt 404 unless @sync_process
+    end
+  end
+end